Restructure project and add CMake testing support
This change restructures the project into a more general c++ project
structure, getting away from the structure Visual Studio projects use.

This change also adds support for testing with the GTest framework via
CMake. Note that the tests have to be actualized with the current source
tree in order for them to run.