bug(#15): Fix message passing to stopped processes
This change fixes a bug where sending a message to a stopped process
(gracefully or forced) would crash the simulation kernel with an
UninitializedPropertyAccessException.

This was caused by the fact that these processes still existed in the
registry, which caused the kernel to lookup a non-existent continuation
of a process.

This change will make the kernel remove stopped processes from the
registry, so they cannot be found anymore.

Closes #15